+++ addons/sale/static/src/js/quotation_view.js
@@ -103,3 +103,37 @@ odoo.define('sale.quotation_view', function (require) {
         });
     });
 });
+
+
+odoo.define('sale.payment_method', function (require) {
+'use strict';
+
+    var ajax = require('web.ajax');
+    var saleQuotePayment = require('payment.payment_method');
+
+    saleQuotePayment.include({
+        payment_transaction_action: function(acquirer_id, params){
+            // override this function as per controllers(route) of module wise
+            if($("#website_quote_payment").length){
+                var href = $(location).attr("href"),
+                    payment_request_id = href.match(/quote\/([0-9]+)/),
+                    access_token = href.match(/quote+\/([^\/?]*)/),
+                    params = {};
+
+                params.access_token = access_token ? access_token[1] : '';
+                params.payment_request_id = payment_request_id ? payment_request_id[1] :
+                ajax.jsonRpc('/quote/transaction/' + acquirer_id, 'call', params).then(function (data) {
+                    $(data).appendTo('body').submit();
+                });
+            }
+            this._super(acquirer_id, params)
+        },
+    });
+
+     $(document).ready(function () {
+        if($("#website_quote_payment").length){
+            var sale_quote_payment = new saleQuotePayment();
+            sale_quote_payment.attachTo($("#website_quote_payment"));
+        }
+    });
+});
